requirements.txt
torch~=1.6.0
torchvision~=0.7.0
tensorboardX~=2.1
tensorboard
tensorflow
easydict~=1.9
pandas~=1.1.3
numpy~=1.18.5
scipy~=1.5.2

Pillow~=8.0.0
opencv-python~=4.4.0.42
matplotlib~=3.3.1
timm~=0.3.2